codebeat fix errors
