+++ bethink/src/components/settings/SecureSettings.js
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import {Text, StyleSheet} from 'react-native';
 import FingerprintScanner from 'react-native-fingerprint-scanner';
-import {ItemSecurity} from './ItemSecurity';
+import {ContainerItemSecurity} from './ContainerItemSecurity';
 
 export const SecureSettings = ({swithSecurity, setSwithSecurity}) => {
   const {pin, biometric} = swithSecurity;
@@ -9,35 +9,18 @@ export const SecureSettings = ({swithSecurity, setSwithSecurity}) => {
 
   const isSensorAvailable = () => {
     FingerprintScanner.isSensorAvailable()
-      .then(biometryType => {
-        setPermission(true);
-      })
-      .catch(error => setPermission(false));
+      .then(biometryType => {setPermission(true);}).catch(error => setPermission(false));
   };
 
   useEffect(() => {
     isSensorAvailable();
   }, []);
-
   return (
     <>
-      <ItemSecurity
-        icon="grid-outline"
-        txt="Pin"
-        enable={pin}
-        setSwithSecurity={setSwithSecurity}
-      />
-      <ItemSecurity
-        icon="finger-print-outline"
-        txt="Biometric"
-        enable={biometric}
-        setSwithSecurity={setSwithSecurity}
-        permission={permission}
+      <ContainerItemSecurity pin={pin} setSwithSecurity={setSwithSecurity} permission={permission} biometric={biometric}
       />
       {!permission && (
-        <Text style={styles.msgPermission}>
-          This option is not enabled on your device
-        </Text>
+        <Text style={styles.msgPermission}>This option is not enabled on your device</Text>
       )}
     </>
   );
