+++ bethink/src/screens/notes/NotesScreen.js
@@ -1,43 +1,27 @@
 import React, {useState} from 'react';
-import {View, Text, StyleSheet, ScrollView} from 'react-native';
+import {View, StyleSheet, ScrollView} from 'react-native';
 import {useSelector} from 'react-redux';
 import {ContainerMain} from '../../components/home/ContainerMain';
 import {ButtonAddNotes} from '../../components/notes/ButtonAddNotes';
-import {CardNotes} from '../../components/notes/CardNotes';
+import {ConditionalNotes} from '../../components/notes/ConditionalNotes';
 import {ModalPinNote} from '../../components/notes/ModalPinNote';
-import {NoNotes} from '../../components/notes/NoNotes';
 import {PriorityNotes} from '../../components/notes/PriorityNotes';
+import { TitleNote } from '../../components/notes/TitleNote';
 
 export const NotesScreen = ({navigation}) => {
-  const {mode} = useSelector(state => state.ui);
   const {notes} = useSelector(state => state.notes);
   const [showModalPin, setShowModalPin] = useState(false);
   return (
     <ContainerMain>
       <ScrollView showsVerticalScrollIndicator={false}>
         <View style={styles.container}>
-          <View style={styles.containerCenter}>
-            <Text style={[styles.title, mode === 'light' && styles.titleLight]}>
-              Notes
-            </Text>
-          </View>
+          <TitleNote />
           <PriorityNotes />
-          {notes.length > 0 ? (
-            <CardNotes
-              notes={notes}
-              navigation={navigation}
-              showModalPin={showModalPin}
-              setShowModalPin={setShowModalPin}
-            />
-          ) : (
-            <NoNotes mode={mode} />
-          )}
+          <ConditionalNotes notes={notes} navigation={navigation} showModalPin={showModalPin} setShowModalPin={setShowModalPin}
+          />
         </View>
       </ScrollView>
-      <ModalPinNote
-        showModalPin={showModalPin}
-        navigation={navigation}
-        setShowModalPin={setShowModalPin}
+      <ModalPinNote showModalPin={showModalPin} navigation={navigation} setShowModalPin={setShowModalPin}
       />
       <ButtonAddNotes navigation={navigation} />
     </ContainerMain>
@@ -47,17 +31,5 @@ export const NotesScreen = ({navigation}) => {
 const styles = StyleSheet.create({
   container: {
     marginHorizontal: 20,
-  },
-  title: {
-    color: 'white',
-    fontWeight: 'bold',
-    fontSize: 25,
-  },
-  titleLight: {
-    color: 'black',
-  },
-  containerCenter: {
-    alignItems: 'center',
-    marginVertical: 20,
   },
 });
